updated developers types
