+++ client/src/app/meetDevelopers/page.tsx
@@ -1,10 +1,9 @@
 import DevelopersGrid from '@/components/DevelopersGrid';
-import { developers } from '@/data/developers';
 
 function App() {
   return (
     <div className="App">
-      <DevelopersGrid developers={developers} />
+      <DevelopersGrid />
     </div>
   );
 }
