+++ client/src/data/developers.json
@@ -1,0 +1,17 @@
+
+{
+  "developers": [
+    {
+      "name": "Mohammad Nafis Fuad",
+      "role": "Frontend Developer",
+      "image": "",
+      "facebook": "www.facebook.com/mn.fuad26"
+    },
+    {
+      "name": "",
+      "role": "",
+      "image": "",
+      "facebook": "#"
+    }
+ ]
+}
