--- client/src/data/developers.ts
@@ -1,16 +1,0 @@
-import { Developer } from '@/types/developer';
-
-export const developers: Developer[] = [
-  {
-    name: "Mohammad Nafis Fuad",
-    role: "Frontend Developer",
-    image: "Image not assigned",
-    facebook: "#"
-  },
-  {
-    name: "Dev Name",
-    role: "Dev Designation",
-    image: "Dev Image",
-    facebook: "#"
-  },
-];
