--- client/src/types/developer.ts
@@ -1,6 +1,0 @@
-export interface Developer {
-  name: string;
-  role: string;
-  image: string;
-  facebook?: string;
-}
